fix item condition oversight
this is a maybe a little bit inefficient because it will always check
all the chains and then the item name itself. I can cut this down, but
each of these calls is O(1) so it's not too bad. I'll change it later
once I'm in a bit more of an optimizationy mood.