replace malformed region check generation for global mode
This commit fixes an issue I noticed where global slot locations would
be considered in logic if you could reach the region connected to the
actual shop region. This works fine except when the shop region is
connected with conditions, at which point you may be required to acquire
an item in Harbor Items + Red Flame with only the requirements for
Harbor Items.

The fix for this, and the way I intended for it to be written in the
first place, is to have it check to see if the shop itself is reachable.